pyproject.toml
[project]
name = "learning-program"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.64.0",
    "datetime>=5.5",
    "duckdb>=1.3.2",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "openai>=1.101.0",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "plotly-express>=0.4.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
